RequestUri methods
